Updated backend

+++ backend/src/main/java/by/bsuir/greenfood/service/impl/BagServiceImpl.java
@@ -2,6 +2,7 @@ package by.bsuir.greenfood.service.impl;
 
 import by.bsuir.greenfood.mapper.BagMapper;
 import by.bsuir.greenfood.model.dto.Bag;
+import by.bsuir.greenfood.model.dto.Dish;
 import by.bsuir.greenfood.model.entity.BagEntity;
 import by.bsuir.greenfood.repo.BagRepository;
 import by.bsuir.greenfood.service.BagService;
@@ -13,6 +14,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 /**
@@ -24,6 +26,8 @@ import java.util.UUID;
 @Service
 @RequiredArgsConstructor
 public class BagServiceImpl implements BagService {
+
+  private static final String BAG_NOT_FOUND_MESSAGE = "Bag not found!";
 
   private final BagMapper mapper;
   private final BagRepository repository;
@@ -51,6 +55,25 @@ public class BagServiceImpl implements BagService {
 
   @Override
   @Transactional
+  public int countBags(UUID userId) {
+    return bagService.findAllByUserId(userId).stream()
+        .map(Bag::getCount)
+        .reduce(0, Integer::sum);
+  }
+
+  @Override
+  @Transactional
+  public Double sumBags(UUID userId) {
+    return bagService.findAllByUserId(userId).stream()
+        .map(bag -> {
+          Dish dish = dishService.getDishById(bag.getDishId());
+          return bag.getCount() * dish.getPrice();
+        })
+        .reduce(0.0, Double::sum);
+  }
+
+  @Override
+  @Transactional
   public List<Bag> findAllByUserId(UUID userId) {
     return repository.findAllByUserId(userId).stream()
         .map(mapper::entityToDto)
@@ -71,7 +94,16 @@ public class BagServiceImpl implements BagService {
   }
 
   @Override
+  @Transactional
   public void deleteBag(UUID userId, UUID dishId) {
-    repository.deleteByUserIdAndDishId(userId, dishId);
+    Bag bag = Optional.ofNullable(bagService.findByUserIdAndDishId(userId, dishId))
+        .orElseThrow(() -> new RuntimeException(BAG_NOT_FOUND_MESSAGE));
+
+    if (bag.getCount() == 1) {
+      repository.deleteByUserIdAndDishId(userId, dishId);
+    } else {
+      bag.setCount(bag.getCount() - 1);
+      repository.save(mapper.dtoToEntity(bag));
+    }
   }
 }
